+++ tests/builder.ml
@@ -193,10 +193,11 @@ let make_points ~rng ~min_year ~max_year_inc ~max_precision =
   | _ -> failwith "Unexpected case"
 
 let make_hms ~rng =
-  Time.Hms'.make_exn
+  Timedesc.Time.make_exn
     ~hour:(rng () mod 24)
     ~minute:(rng () mod 60)
     ~second:(rng () mod 60)
+    ()
 
 let make_hms_intervals_inc ~rng =
   Time.hms_intervals ~inc_exc:`Inc (make_hms ~rng) (make_hms ~rng)